refactor(ide): name activity rail as panel
