+++ dashboard/design-system/CHANGELOG.md
@@ -57,7 +57,7 @@ Stage: legacy alias cleanup + KeeperBadge migration completion + token codificat
 
 - **Run activity store substrate**
   - `src/components/ide/run-activity-store.ts` adds a typed ACTIVITY THIS RUN event store with run scoping, newest-first ordering, keeper grouping, and capped visible history.
-  - `src/components/ide/ide-activity-mock.ts` now consumes the store instead of embedding activity ordering and keeper hue mapping directly in the renderer.
+  - `src/components/ide/ide-activity-panel.ts` now consumes the store instead of embedding activity ordering and keeper hue mapping directly in the renderer.
 
 - **IDE INTERJECT store substrate**
   - `src/components/ide/interject-store.ts` adds typed active-keeper/message state for the INTERJECT rail and explicit availability reasons for Send/Approve/Pause/Drain actions.
