+++ dashboard/src/components/ide/ide-activity-panel.a11y.test.ts
@@ -1,14 +1,20 @@
 // @vitest-environment happy-dom
-import { afterEach, beforeEach, describe, expect, it } from 'vitest'
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { render } from 'preact'
 import { html } from 'htm/preact'
 import { axe } from 'jest-axe'
-import { IdeActivityMock } from './ide-activity-mock'
+import { IdeActivityPanel } from './ide-activity-panel'
 
-describe('IdeActivityMock a11y', () => {
+describe('IdeActivityPanel a11y', () => {
   let container: HTMLElement
 
   beforeEach(() => {
+    vi.stubGlobal('fetch', vi.fn(async () =>
+      new Response(JSON.stringify({ events: [] }), {
+        status: 200,
+        headers: { 'Content-Type': 'application/json' },
+      }),
+    ))
     container = document.createElement('div')
     document.body.appendChild(container)
   })
@@ -16,10 +22,11 @@ describe('IdeActivityMock a11y', () => {
   afterEach(() => {
     render(null, container)
     document.body.removeChild(container)
+    vi.unstubAllGlobals()
   })
 
   it('renders the run activity pane accessibly', async () => {
-    render(html`<${IdeActivityMock} />`, container)
+    render(html`<${IdeActivityPanel} />`, container)
     expect(await axe(container)).toHaveNoViolations()
   })
 })
